requirements.txt
-i https://pypi.org/simple
discord.py==2.2.2
uvicorn==0.20.0
requests==2.28.2
